REFACTOR pyinstalller_runtime, code cleanup
